--- packages/test-app/app/modifiers/autofocus.js
@@ -1,6 +1,0 @@
-import { modifier } from 'ember-modifier';
-
-export default modifier(function autofocus(element) {
-  const childElement = element.querySelector('input');
-  childElement.focus();
-});
